% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dafrjulia-package.R
\docType{package}
\name{dafrjulia-package}
\alias{dafrjulia}
\alias{dafrjulia-package}
\title{dafrjulia: R Interface to the Julia 'DataAxesFormats' Package}
\description{
An R interface to 'DataAxesFormats.jl' \url{https://github.com/tanaylab/DataAxesFormats.jl}, a 'Julia' \url{https://julialang.org/} package providing a uniform interface for accessing multi-dimensional data arranged along arbitrary axes. This package generalizes and extends 'AnnData'-like functionality, supporting both 1D and 2D data, multiple data modalities, and efficient memory management. 'dafrjulia' enables R users to create, read, modify, and analyze complex biological datasets such as single-cell data with multiple modalities (RNA-seq, ATAC-seq, etc.). The package implements efficient memory management strategies and provides a seamless interface between R and 'Julia' for handling large-scale biological data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tanaylab.github.io/dafrjulia/}
  \item \url{https://github.com/tanaylab/dafrjulia}
  \item Report bugs at \url{https://github.com/tanaylab/dafrjulia/issues}
}

}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il} (\href{https://orcid.org/0000-0002-8458-9507}{ORCID})

Authors:
\itemize{
  \item Oren Ben-Kiki \email{oren@ben-kiki.org}
}

Other contributors:
\itemize{
  \item Weizmann Institute of Science [copyright holder]
}

}
\keyword{internal}
