% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{files_to_zarr}
\alias{files_to_zarr}
\title{Convert a native files Daf to Zarr format}
\usage{
files_to_zarr(files_path, zarr_path)
}
\arguments{
\item{files_path}{Path to the source native files Daf directory}

\item{zarr_path}{Path to the destination \code{.daf.zarr} (created)}
}
\value{
NULL, invisibly. Called for its side effect.
}
\description{
Convert a native files Daf to Zarr format
}
