% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_empty_dense_vector}
\alias{get_empty_dense_vector}
\title{Get an empty dense vector for filling}
\usage{
get_empty_dense_vector(daf, axis, name, eltype, overwrite = FALSE)
}
\arguments{
\item{daf}{A Daf object}

\item{axis}{Name of the axis}

\item{name}{Name of the vector property}

\item{eltype}{Element type for the vector (e.g., "Float64", "Int32")}

\item{overwrite}{Whether to overwrite if vector already exists (FALSE by default)}
}
\value{
A Julia vector object backed by the (uninitialized) property storage
}
\description{
Returns an empty dense vector for the specified axis and property, which can be filled
in-place. This is useful for efficiently constructing large vectors without allocating
temporary storage. After filling, the vector is automatically stored in the Daf object.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/writers.html#DataAxesFormats.Writers.get_empty_dense_vector!}{documentation} for details.
}
