% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder_axes}
\alias{reorder_axes}
\title{Reorder the entries of one or more axes}
\usage{
reorder_axes(daf, permutations)
}
\arguments{
\item{daf}{A writable Daf object, or a list of writable Daf objects.}

\item{permutations}{A named list mapping each axis name to a 1-based integer
permutation vector of that axis's entries.}
}
\value{
NULL, invisibly. Called for its side effect.
}
\description{
Permute the entries of the named axes in a writable leaf Daf (or a set of leaf
Dafs sharing those axes), rewriting all data that depends on them. Not valid for
chain or view Dafs. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.3.0/reorder.html}{documentation} for details.
}
