% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reset_reorder_axes}
\alias{reset_reorder_axes}
\title{Roll back an interrupted axis reorder}
\usage{
reset_reorder_axes(daf)
}
\arguments{
\item{daf}{A writable Daf object, or a list of writable Daf objects.}
}
\value{
TRUE if a stale reorder was rolled back, FALSE if there was nothing to do.
}
\description{
If a previous \code{reorder_axes} call was interrupted, restore the Daf(s) to their
pre-reorder state. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.3.0/reorder.html}{documentation} for details.
}
