% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{zarr_to_files}
\alias{zarr_to_files}
\title{Convert a Zarr Daf to native files format}
\usage{
zarr_to_files(zarr_path, files_path)
}
\arguments{
\item{zarr_path}{Path to the source \code{.daf.zarr}}

\item{files_path}{Path to the destination native files Daf directory (created)}
}
\value{
NULL, invisibly. Called for its side effect.
}
\description{
Convert a Zarr Daf to native files format
}
